requirements.txt
Django==5.2.4
djangorestframework
django-filter
django-cors-headers
Pillow
psycopg2-binary 
gunicorn 
whitenoise
drf-spectacular
drf-spectacular[sidecar]
python-slugify
djangorestframework-simplejwt
requests
cloudinary 
django-cloudinary-storage
google-genai
